entity va service cart publish

+++ src/main/java/com/mangaone/repository/MangaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MangaRepository extends JpaRepository<Manga, Integer> {
+public interface MangaRepository extends JpaRepository<Manga, Long> {
     // Để trống như này là đủ, JpaRepository sẽ tự lo hết các hàm tìm kiếm, lưu, xóa.
 }
