+++ src/main/java/com/mangaone/service/MangaService.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public interface MangaService {
     List<Manga> getAllMangas(); // Liệt kê tất cả truyện
-    Manga getMangaById(Integer id); // Xem chi tiết 1 cuốn
+    Manga getMangaById(Long id); // Xem chi tiết 1 cuốn
     void saveManga(Manga manga); // Lưu hoặc cập nhật truyện
-    void deleteManga(Integer id); // Xóa truyện
+    void deleteManga(Long id); // Xóa truyện
 }
